updated TweetDTO

+++ src/main/java/com/github/yvasyliev/model/dto/TweetDTO.java
@@ -2,5 +2,5 @@ package com.github.yvasyliev.model.dto;
 
 import java.time.LocalDateTime;
 
-public record TweetDTO(Long id, String text, LocalDateTime createdAt, Long replies, Long likes) {
+public record TweetDTO(Long id, String text, LocalDateTime createdAt, Integer replies, Integer likes) {
 }
